% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.cv_rpair.R
\name{coef.cv_rpair}
\alias{coef.cv_rpair}
\title{Extract coefficients from a cv_rpair object}
\usage{
\method{coef}{cv_rpair}(object, s = c("lambda.1se", "lambda.min"))
}
\arguments{
\item{object}{A fitted rpair object.}

\item{s}{Value(s) of the penalty parameter lambda at which predictions are required. Default is the value
s="lambda.1se".}
}
\value{
A matrix of coefficients for the provided lambda.
}
\description{
This function is equivalent to calling predict(object, type="coefficients") on a cv_rpair object.
}
\examples{
efit = cv_rpair(ds1_x, ds1_y, nlambda=25, loss_type="exp")
ec = coef(efit, s="lambda.1se")

}
