% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cv_rpair.R
\name{predict.cv_rpair}
\alias{predict.cv_rpair}
\title{Predict or extract values from cv_rpair object}
\usage{
\method{predict}{cv_rpair}(object, newx, s = c("lambda.1se", "lambda.min"), ...)
}
\arguments{
\item{object}{A fitted cv_rpair object.}

\item{newx}{A matrix of new values. This argument is not required for type=c("coefficients", "nonzero").}

\item{s}{Value(s) of the penalty parameter lambda at which predictions are required. Default is the value
s="lambda.1se".}

\item{\dots}{Additional parameters to pass to the predict function.}
}
\value{
The object returned depends on the values passed via ... . See predict.rpair for details.
}
\description{
Predicts fitted values from a cross-validated rapir object given a matrix of new values. Can
also extract and return coefficient and nonzero lists. Function internally calls predict.rpair
on the cv_rpair master fit object.
}
\author{
KC
}
