% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.rpair.R
\name{coef.rpair}
\alias{coef.rpair}
\title{Extract coefficients from an rpair object}
\usage{
\method{coef}{rpair}(object, s = NULL)
}
\arguments{
\item{object}{A fitted rpair object.}

\item{s}{Value(s) of the penalty parameter lambda at which predictions are required. Default is the entire
sequence used to create the model.}
}
\value{
A matrix of coefficients.
}
\description{
This function is equivalent to calling predict(object, type="coefficients").
}
\examples{
efit = rpair_gloss(surv_x, surv_cp, pmax = 50, loss_type = "exp")
coefmat = coef(efit)

}
