% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_rpair.R
\name{cv_rpair}
\alias{cv_rpair}
\title{Cross-Validation function for rpair_gloss}
\usage{
cv_rpair(
  x,
  y,
  lambda = NULL,
  nlambda = 100,
  type.measure = c("deviance", "cindex"),
  nfolds = 10,
  foldid = NULL,
  alignment = c("lambda", "fraction"),
  grouped = FALSE,
  keep = FALSE,
  ...
)
}
\arguments{
\item{x}{Input matrix as in rpair_gloss.}

\item{y}{Response as in rpair_gloss.}

\item{nlambda}{Number of lambda values. Default: 100.}

\item{type.measure}{Loss to use for cross-validation. Available loss functions are "deviance" and "cindex".}

\item{nfolds}{Number of folds. Default: 10.}

\item{alignment}{Determines}

\item{grouped}{Not implemented in this version. Default: FALSE.}

\item{keep}{If keep=TRUE, returns a list of fitted values for each fold and a corresponding list of foldids.
Default: FALSE.}

\item{\dots}{Additional parameters to pass to the rpair_gloss function.}
}
\value{
An object of class "cv_rpair" containing the following list of values:
}
\description{
Performs k-fold cross-validation for rpair_gloss.
}
