% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_rpair.R
\name{cv_rpair}
\alias{cv_rpair}
\title{Cross-Validation function for rpair_gloss and rpair_hinge.}
\usage{
cv_rpair(
  x,
  y,
  loss_type = c("exp", "log", "sqh", "huh"),
  lambda = NULL,
  nlambda = 100,
  type.measure = c("deviance", "cindex"),
  nfolds = 5,
  foldid = NULL,
  alignment = c("fraction", "lambda"),
  grouped = FALSE,
  keep = FALSE,
  ...
)
}
\arguments{
\item{x}{Input matrix as in rpair_gloss and rpair_hinge.}

\item{y}{Response as in rpair_gloss and rpair_hinge.}

\item{loss_type}{Loss function to use. One of c("exp", "log" ,"sqh", "huh"). Default: "exp".}

\item{nlambda}{Number of lambda values. Default: 100.}

\item{type.measure}{Loss to use for cross-validation. Available loss functions are "deviance" and "cindex".}

\item{nfolds}{Number of folds. Default: 5.}

\item{foldid}{Optional vector of values between 1 and nfolds identifying which fold each observation should be in.
If provided, nofolds can be missing. Pairs caanot be provided as input when foldid is missing.}

\item{alignment}{Determines whether the predictions in each fold are aligned according to the fraction
of progress along each fold ("fraction", default) or to use the lambda values computed on the master fit
to line up the data on each fold ("lambda"). See \link[glmnet]{cv.glmnet} for details. One of
c("fraction", "lambda"). Default: "fraction".}

\item{grouped}{Not implemented in this version. Default: FALSE.}

\item{keep}{If keep=TRUE, returns a list of fitted values for each fold and a corresponding list of foldids.
Default: FALSE.}

\item{\dots}{Additional parameters to pass to the rpair_gloss or rpair_hinge function.}
}
\value{
An object of class "cv_rpair" containing the following list of values:
}
\description{
Performs k-fold cross-validation for rpair_gloss or rpair_hinge.
}
