% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.rpair.R
\name{predict.rpair}
\alias{predict.rpair}
\title{Predict or extract values from rpair object}
\usage{
\method{predict}{rpair}(object, newx, s = NULL, type = c("link", "coefficients", "nonzero"))
}
\arguments{
\item{object}{A fitted rpair object.}

\item{newx}{A matrix of new values. This argument is not required for type=c("coefficients", "nonzero").}

\item{s}{Value(s) of the penalty parameter lambda at which predictions are required. Default is the entire
sequence used to create the model.}

\item{type}{Type of prediction desired. Type "link" gives linear predictors, type "coefficients" returns a matrix
of coefficients, and type "nonzero" returns a list of the indices of the nonzero coefficients.}
}
\value{
The object returned depends on the value of \code{type}. See \code{type}.
}
\description{
Predicts fitted values given a matrix of new values. Can also extract and return coefficient and nonzero lists.
}
\examples{
efit = rpair_gloss(rpair::ds1_x, rpair::ds1_y, standardize = F, pmax = 50, loss_type = "exp")
newpred = predict(efit, rpair::ds1_x, type="link")
nonzero = predict(efit, type = "nonzero")

}
\author{
KC
}
