% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpair.R
\name{rpair}
\alias{rpair}
\title{Fit GLM or SVM for pairwise survival ranking models}
\usage{
rpair(
  x,
  y,
  loss_type = c("exp", "log", "sqh", "huh"),
  alpha,
  nlambda,
  lambda.min.ratio,
  lambda,
  lambda2,
  standardize,
  thresh,
  dfmax,
  pmax,
  penalty.factor,
  pf1,
  pf2,
  maxit,
  type.logistic,
  delta
)
}
\arguments{
\item{x}{Input matrix of dimension nobs x nvars; each row is an observation vector.}

\item{y}{Pairwise ranking analysis response variable. This function supports three types of input types:
(1) continuous values, (2) survival data, and (3) ranked pairs.}

\item{loss_type}{Loss function to use. Equivalent to "family" in glmnet package. One of c("log", "exp").}

\item{alpha}{For GLM models only. The elasticnet mixing parameter. See \link[glmnet]{glmnet} for details.
Default: 1.0 (lasso penalty).}

\item{nlambda}{Number of lambda values to evaluate. Default: 100.}

\item{lambda.min.ratio}{For SVM models, this is equivalent to lambda.factor. Smallest value for lambda as a
fraction of lambda.max, the (data derived) entry value. See \link[glmnet]{glmnet} or \link[gcdnet]{gcdnet} for
details. Default: ifelse(nobs0 <= nvars, 1e-2, 1e-4)}

\item{lambda}{A user supplied lambda sequence. Overrides the typical usage in which a lambda sequence is computed
using nlambda and lambda.min.ratio. Provide a decreasing sequence of lambda values with at least 2 entries.
Default: NULL.}

\item{lambda2}{For SVM models only. Regularization parameter lambda2 for the quadratic penalty of the coefficients.
Default: 0.}

\item{standardize}{Logical flag for x variable standardization. Default: FALSE.}

\item{thresh}{For SVM models, this is equivalent to eps. Convergence threshold for coordinate descent.
GLM Default: 1e-7, SVM Default: 1e-6.}

\item{dfmax}{Limit the maximum number of variables in the model. Default: nvars+1.}

\item{pmax}{Limit the maximum number of variables that can be nonzero. Default: min(dfmax*2+20, nvars).}

\item{penalty.factor}{For GLM models only. Vector of penalty factors to apply to each coefficient.
Default: rep(1, nvars).}

\item{pf2}{For SVM models only. L2 penalty factor of length p used for adaptive LASSO or adaptive elastic net. See
\link[gcdnet]{gcdnet} for details. Default: rep(1, nvars).}

\item{maxit}{Maximum number of passes over the data for all lambda values. Default: 100000.}

\item{type.logistic}{For GLM models only. One of c("Newton", "modified.Newton"). If "Newton" then the exact
hessian is used, while "modified.Newton" uses an upper-bound on the hessian and can be faster.
Default: "Newton".}

\item{delta}{For SVM models only. The parameter delta in the HHSVM model. Must be greater than 0. Default: 3.}

\item{pf}{For SVM models only. L1 penalty factor of length p used for adaptive LASSO or adaptive elastic net. See
\link[gcdnet]{gcdnet} for details. Default: rep(1, nvars).}
}
\value{

}
\description{
Fit a generalized linear model or SVM model for pairwise survival ranking models. GLM models use either logistic
loss (concordance regression) or exponential loss. SVM models use squared hinge loss or huberized hinge loss.
Uses a modified implementation of \link[glmnet]{glmnet} for logistic and exponential loss and a modified
implementation of \link[gcdnet]{gcdnet} for squared and huberized hinge loss. Refer to glmnet and gcdnet
documentation for further details.
}
\examples{
# GLM
efit = rpair(surv_x, surv_cp, standardize = F, pmax = 50, loss_type = "exp")
# SVM
sfit = rpair(surv_x, surv_cp, standardize = F, pmax = 50, loss_type = "sqh")

}
\author{
mubu, KC
}
